django-backend/requirements.txt
drf-yasg==1.21.8
Django==5.1.4
djangorestframework==3.15.2
pytz==2024.2